+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Changed
 
 - Revised export of Docker images, allowing to easily create GitHub releases with all images
+- Renamed analysis to solving
 
 ### Removed